bot/requirements.txt
python-telegram-bot[ext]==21.6
httpx>=0.27
pydantic-settings>=2.0
